+++ bin/build.sh
@@ -1,9 +1,12 @@
 #!/usr/bin/env bash
 
+# Ensure the build fails if TypeScript fails
 set -e
 
+# Lint TypeScript source code
 npx tsc --noEmit --pretty
 
+# Bundle into JavaScript
 npx esbuild \
   --bundle \
   --format=cjs \