+++ src/Scripts/commands/all.ts
@@ -1,1 +1,0 @@
-export { reloadCommand } from "./reload-command";